[QTI-152] Add transport to enos DSL and module generator
This change adds support for `transport` stanzas in the root Enos DSL to
allow authors to create transport configurations that can be reused.

`scenario` blocks now also support a `transport` attribute that, when
present, will generate a provider level transport configuration for the
generated module.

If "default" labeled `transport` is defined it will automatically be
inherited by `scenario`'s which have no specified a transport.

As part of this change we also reorganized the CLI "integration" tests
into a top-level `acceptance` package that executes the binary. We also
added some new acceptance helpers in preperation for a full AWS
end-to-end test.

Signed-off-by: Ryan Cragun <me@ryan.ec>