Updated call to validate workflow and added comments
Updated the call in pr_build.yml to the validate workflow to reference the main branch rather than the prior PR branch. Added comments explaining that anytime a change is being made to validate.yml, we will need to temporarily change this reference to the active PR branch, otherwise GitHub Actions will be looking for the version of validate.yml that's currently on main. It should be changed back to main after being merged.