[QTI-165] Add step variable support
In order to pass information from one step module to another,
we need to be able to pipe the outputs from one generated
Terraform module to another. This change achieves that by
decoding "step" variables as a capsule type and either
saves it as a traversal to a module or the known value.

* Add StepVariable capsule type
* Add tests for decoding the capsule type
* Update scenario step decoding to use the new capsule type
* Add support for validating step variables with references to the
  decoder helpers.
* Add step references to acceptance generation test
* Add generator support for step var types

Signed-off-by: Ryan Cragun <me@ryan.ec>