[QTI-253] Refactor CLI into client and server
In order to better multiplex multiple concurrent scenario operations,
keep and serialize scenario state and metadata, and build a richer UX
experience, we needed to re-architect Enos into a client/server model.

We needed to do this refactor to build a consistent interface between
scenarios and the CLI and it also adds future-proofing where Enos is
able to run as a daemonized server.

We decided to go with gRPC and Protobuf as RPC and message descriptors.
We chose this because Protobuf allows guarantees on forward and backwards
compatibility, serialization as text or JSON, and has an efficient wire
format that we can use in a client/server model. gRPC allows us to
generate efficient boilerplate using our message descriptors.

As the change is quite large due to the refactor, it's important to note
what material changes happen:

* Enos command output is drastically different. Rather than piping
  Terraform output directly, it now displays them as events for the
  scenario. Further work in future PR's will improve this even more.
* Enos is able to output JSON machine output for all commands. We use
  json field options to normalize our server responses as JSON.
* Except for `output`, `exec`, and `destroy`, all other scenario commands
  will now execute their dependencies. I.e. you no longer need to run
  the `validate` command before the `launch` command.

List of changes:

* Define an initial Protobuf schema for Enos
* Add Protobuf compilation, formatting, and linting with Buf
* Refactor Makefile
* Refactor enos commands into a gRPC requests
* Add support for passing Terraform output through the executor
* Add CLI human and machine views that utilize the request responses
* Refactor acceptance tests to use machine output
* Update Go deps

Signed-off-by: Ryan Cragun <me@ryan.ec>