[QTI-239] Add fmt command
This change introduces the `fmt` command to enos. This allows scenario
authors to have built-in way to normalize their configuration and
variables files. By default the fmt command will look in the given
working directory for enos files and will normalize their formatting.
This behavior can be overridden by passing a path to the command or - to
read from STDIN.

If authors instead prefer checking they can pass the `--check` flag
which disables writing and instead will list files that would change and
exit with 3 if changes could be made. This is useful for automated CI
checking.

Authors may also choose to pass the --diff flag, which produces a
unified diff of the changes and writes the to STDOUT.

* Move status package
* Clean up flag descriptions
* Create a fmt sub-command
* Add acceptance test

Signed-off-by: Ryan Cragun <me@ryan.ec>