[QTI-254] Remove auto depends_on from steps (#52)
* [QTI-254] Remove auto depends_on from steps

In Terraform, we've found that automatically depending on prior
modules can have unintended consequences. When a module `depends_on`
another module output that has a collection it can require Terraform
to plan in multiple phases which can break modules that output dynamic
collections.

We had to make a choice: keep auto-`depends_on` as a feature in Enos but
allow scenario authors an escape hatch that would disable depending on
prior steps, or remove auto-`depends_on` and allow Terraform to handle
dependencies as it normally would. The former is simple but would
require scenario authors to interpret strange Terraform error messages.
The latter could run into problems where a step might have an implicit
dependency on a prior step that Terraform can't reason about. In those
cases the author would need to specify a `depends_on` for that step.
It's a bit more complicated but maps to how Terraform actually works. We
chose the latter option.

* Remove auto-`depends_on` feature.
* Allow scenario authors to configure `depends_on` values for steps.

Signed-off-by: Ryan Cragun <me@ryan.ec>