[QTI-257] Allow customized handling of warnings (#59)
This change modifies the default behavior of Enos to allow warnings
severity diagnostics by default. When a warning diagnostic is
encountered, we'll render warning UI messages as appropriate but allow
execution of scenario operations to proceed. To disable this behavior,
a user may now pass the `--fail-on-warnings` flag which will prevent
scenario operations from processing new tasks when a warning diagnostic
has been generated by either Enos or Terraform. It was also exit with a
failure if a warning diagnostic is returned.

* Update root command description
* Rename `client-log-level` flag to `log-level`
* Add `fail-on-warnings` flag to the root command and proto schema
* Add support for optionally continuing or failing on warning
  diagnostics in the Terraform executor
* Clean up server helpers to only return responses with diagnostics
* Pass generator diagnostics through all server endpoints that
  generate modules
* Add support for optionally continuing or failing on warning
  diagnostics in the UI status package
* Add warning diagnostic rendering to basic text views
* Add acceptance test for validation behavior for both default and
  fail-on-warning cases

Signed-off-by: Ryan Cragun <me@ryan.ec>