[QTI-261] Don't alias default providers (#60)
This change was intended to allow a small quality of life feature for
scenario authors. Our previous implementation of the providers decoder
and transpiler would always add alias information for each defined
provider based on its second name label. In practice this allows an
author to define lots of addressable providers so that they can use the
correct one in their scenario. A result is that there will never be a
"default" provider in the rendered module, e.g. no provider that does
not have an alias. That means that _every step_ that uses the provider
has to specify it in the providers attribute map.

This change allows a user to specify "default" providers for a provider
type by making its name label "default". When this occurs and it is used
in a scenario we'll automatically render it in the transpiled module
without and alias. This will allow step modules to inherit it by
default. As a side effect we also disallow through a warning diagnostic
the usage of default providers in step attribute maps, as it would be
impossible to assign an un-aliased provider in that configuration.

A side effect of this change was realizing that we were not displaying
decoding diagnostics effectively in the basic text view. Because of
this, decoding is now a first class concept in Enos and the server and
UI now understand which diagnostics are generated during the decoding
phase and that they are displayed only a single time.

* Disallow referring to "default" providers in step attributes
* Don't render "default" providers with an alias
* Rename Scenario.Command in proto schema to Scenario.Operation
* Add decoding as a scenario operation
* Migrate all prior usage of gRPC status and CLI errors to use decoding
  diagnostics when appropriate
* Fix a few missed cases where we weren't considering warning
  diagnostics during decoding and generation

Signed-off-by: Ryan Cragun <me@ryan.ec>