[QT-140] Add initial multiplexing support (#67)
* [QTI-140] Add initial multiplexing support

This commit restructures the way the Enos server dispatches scenario
operations, tracks their state, and reports information to the client.
These changes allow Enos to run multiple scenario operations in parallel
with complete Terraform isolation. They also allow us to provide a rich
near real-time CLI UX.

The server is now responsible for decoding the flight plan, filtering which
scenarios to use, and dispatching an operation for each one. Scenario
server endpoints no longer wait until the completion of operations, they
simply request them and return references to the in-flight operations.
The client can now use the operation status and stream endpoints to
stream near-real-time events, see the operation event history, and get
the operation status.

In practice, this means that execution of various tasks is now highly
generalized. The server expects a request that contains the desired
operation type and the scenario identifiers and executes a worker
function that is responsible for executing the desired task and updating
the server with realtime events.

This change allows us to execute multiple operations in parallel with complete
Terraform isolation with event based near real-time updates. The server
creates a pool of workers which read from the queue and process them as they
come. While most operations will likely execute in the lexicographical order
of their scenario identifiers, no execution order is guaranteed. The
client and various scenario subcommands can configure the worker count
using its associated flag.

As part of this work we now keep track of the operation status and an
event history for every operation that is dispatched to the server. The
current state implementation uses memory so any event history or status
of operations outside of their Terraform state is not persisted. Further
work to serialize state locally for use between Enos invocations will be
required.

* Update to Go 1.19
* Restructure the server executor as an operation runner and refactor
  scenario commands and Terraform commands into runner operations and
  worker functions.
* Add logging to the server
* Decode and filter scenarios in the server.
* Add an operator to handle operations in the server. The operator
  handles dispatch, tracking, status, and event streaming for running
  operations.
* Add in-memory state tracking of operations
* Add operation event handling. The server maintains and event publisher
  which tracks listeners and publishes worker events.
* Rewrite scenario commands to dispatch events, stream events, and
  return a status based on all operation responses.
* Restructure tests to handle new proto operation response types
* Rename the validate command to client and add an alias
* Add error logging to the client
* Add trace level logging
* Create a client operation event streamer

Signed-off-by: Ryan Cragun <me@ryan.ec>

* feedback round 1

Signed-off-by: Ryan Cragun <me@ryan.ec>

* simplify and fix diagnostics status

Signed-off-by: Ryan Cragun <me@ryan.ec>

* review feedback round two

Signed-off-by: Ryan Cragun <me@ryan.ec>

* remove deadcode

Signed-off-by: Ryan Cragun <me@ryan.ec>

Signed-off-by: Ryan Cragun <me@ryan.ec>