[QT-326] add state and deps handling for destroy (#69)
When executing the `scenario destroy` sub-command we will now initialize
the scenario Terraform module before we destroy it. This ensures that
all of the requisite dependencies necessary to create a destroy plan are
available when executing the destroy plan. We also check the existing state
of the module to determine if it has any resources. If no resources are present
in the state there's nothing to destroy, thus we can return early if necessary.
This also has the benefit of allowing us to "destroy" modules that can destroy
until they've reached all requisite targets needed for a complete plan.

* Add `show` sub-operation to destroy operation to read the module state
* Add `init` sub-operation to destroy operation
* Use the module state to determine whether or not a module needs to be
  deleted. This prevents us from trying to delete modules that don't
  have delete-able state.
* Bump the version
* Exit with an error if no scenarios are found matching an operation
  filter

Signed-off-by: Ryan Cragun <me@ryan.ec>