[QT-286] Reconfigure the backend state by default
When configuring Terraform to use a different backend, one must
initialize the module with the `-reconfigure` flag to signal to
Terraform that it can "forget" the prior configuration or the default
configuration of a local backend. In the context of scenarios and
emphemeral infrastructure, this isn't likely to be a common scenario.

This change now passes the `--reconfigure` flag for all `terraform init`
executions by default, thus ensuring that init should always use what is
currently defined in the scenario. If for some reason the author really
wants to execute without modifying the backend they can opt-out to the
default behavior with the `--no-reconfigure` flag.

Signed-off-by: Ryan Cragun <me@ryan.ec>