[QT-529] Always generate a Terraform configuration files
Enos scenarios always override the CLI configuration file to ensure that
we don't accidentally inherit global configuration. Previously we relied
on simply setting the environment variable and geneating actual contents
only if they were set in the scenario. Terraform 1.4.2 changed the
behavior to warn if the environment variable was set to a non-existent
file. Now we always create the file even if it's blank.

Signed-off-by: Ryan Cragun <me@ryan.ec>