[QT-601] Nest scenario step variables in eval context
Previously we'd expose any know scenario step variables at the top-level
of a steps entry in the eval context. This leads to problems if a
Terraform module has both an input and output with the same name.
Consider the following scenario:

```hcl
module "cluster" {
  source = "./modules/cluster"
}

module "worker" {
  source = "./modules/worker"
}

variable "addr" {
  type = string
  default = "http://192.168.0.1"
}

scenario "boundary" {
  step "cluster" {
    module = module.cluster
    variables {
      addr = var.addr
    }
  }

  step "worker" {
    module = module.worker
    variables {
      upstream_addr = step.cluster.addr
    }
  }

  step "worker_downstream" {
    module = module.worker
    variables {
      upstream_addr = step.worker.upstream_addr
    }
  }
}
```

We expect expect `step.cluster.addr` to be the known value of of
`var.addr`. But what happens with step worker here is the problem.
`step.worker.upstream_addr` should be module call to `step.cluster.addr`
but because we've exposed `step.cluster.addr` in the eval context we
don't do a module call, and instead inherit the value from `var.addr`.
The same happens with `step.worker_downstream.upstream_addr`. Instead of
being a module call to `step.worker.upstream_addr` we inherit the know
value of `var.addr`.

With this change that behavior is now default. If you wish to use the
prior behavior (getting a steps known variables) you can access them
with the `variables` key, e.g. `step.cluster.variables.addr`.

- [x] The commit message includes an explanation of the changes
- [x] Manual validation of the changes have been performed (if possible)
- [x] New or modified code has requisite test coverage (if possible)
- [x] I have performed a self-review of the changes
- [x] I have made necessary changes and/or pull requests for documentation
- [x] I have written useful comments in the code

Signed-off-by: Ryan Cragun <me@ryan.ec>