flightplan: fix various potential nil panics
When testing scenario filters with negations for variants that didn't
exist for some scenarios, I ran into a panic in the Matrix implementation
due to not checking for nil and handling it appropriately in our
exclusions. After fixing it I decided to use the `nilaway`[0] static
analysis tool to verify the entire `flightplan` package.

The tool itself definitely has a few rough edges, and there were times
when it was confused and I had to do things more than once in a function
to please it, but overall it found several useful improvements.

It was fairly time consuming so I scoped this to just the `flightplan`
package for now. Eventually it might be nice to enable for the entire
program but we'd either need to architect away our use of global
variables for the UI, server client, and server, or we'd need them to
reintroduce per line `//nolint` directives[1].

It also has a few show stopper bugs like not being able to handle
listening on a nil channel in a select statement.[2] A tool with a
promising and useful future but needs more time for us to enforce.

[0] https://github.com/uber-go/nilaway
[1] https://github.com/uber-go/nilaway/issues/126
[2] https://github.com/uber-go/nilaway/issues/143

Signed-off-by: Ryan Cragun <me@ryan.ec>