buf: upgrade to latest buf which doesn't support our old generate structure
Upgrade to the latest version of buf and with it the updated
configuration formats.

I couldn't find a way for Buf's new CLI to allow us to generate code into
the same directory as the protos like we were doing before without
nesting it a billion directories. As such, we'll move the package out
and and rename the imports.

Signed-off-by: Ryan Cragun <me@ryan.ec>