matrix: don't overwrite matrix eval context on attr only matrices
Matrix blocks are used in several places and depending on the context we
we need slightly different behavior when decoding them.

When we're decoding a "full matrix", i.e. one that must have variants and
may have `include` and `exclude` sub-blocks, we want to first decode the
attributes and update the `matrix` eval context so that it's available
when decoding the next attribute or the the later blocks.

When we're decoding `include` or exclude `blocks`, they themselves are
also decoded as an attrubute only matrix, but we don't want to update
the `matrix` eval context and instead want to inherit the parent
matrix's eval context.

This change makes that so, fixing a bug whereby the `matrix` block only
included the parents eval context on the first `include` or `exclude`
block.

We also end up pinning the version of `random` in our acceptance tests
because a newly released version broke the test output that we expect.
Since we only care about getting a diagnostic warning it should be fine
to pin to that version.

Signed-off-by: Ryan Cragun <me@ryan.ec>