funcs: add some missing terraform funcs
While working on scenario I wanted to use `strcontains()` and discovered
it was a Terraform specific func that isn't part of the `cty` standard
library funcs that we've added to Enos.

This adds some missing `yaml` and Terraform funcs that might be useful:

- `abs()`
- `alltrue()`
- `anytrue()`
- `base64decode()`
- `base64encode()`
- `base64gzip()`
- `endswith()`
- `matchkeys()`
- `one()`
- `startswith()`
- `strcontains()`
- `sum()`
- `textdecodebase64()`
- `textencodebase64()`
- `timestamp()`
- `timecmp()`
- `transpose()`
- `urlencode()`
- `yamldecode()`
- `yamlencode()`

We also update our Go modules as we needed a few new dependencies.

Signed-off-by: Ryan Cragun <me@ryan.ec>