[VAULT-28010] fix(funcs): fix relative paths in `file()`
The recently added `file()` function supports relative paths but they’ll
always be relative to the working dir that exec’s enos. This is usually
fine unless enos is invoked with `--chdir`, in which case we need funcs
that support relative paths to be relative to the `--chdir` working dir.

* Make `file()` support the `--chdir` base path when given a base path
* Bump the version
* Add tests for funcs that support relative paths
* Remove test data that wasn't being used

Signed-off-by: Ryan Cragun <me@ryan.ec>