terraform_cli: support `path` and `env` and in `terraform_cli` blocks
At some point during the serial to multiplex rewrite of Enos, we stopped
respecting the special `path` and `env` attributes in `terraform_cli`
blocks.

I discovered this when attempting to test an older version of Terraform
with `enos` by setting the `path` and it was still resolving the latest
version which it found via the `PATH`.

This resolves that by configuring the Terraform executor in the
workspace with any special `terraform_cli` variables that have been set
in the flight plan.

* Configure the Terraform executor with the scenario "terraform_cli"
  attributes if necessary
* Regenerate protobuf with latest generator
* Update Go modules
* Bump version

Signed-off-by: Ryan Cragun <me@ryan.ec>