VAULT-41529: fix filtering by variant and no scenario name
When we rewrote scenario filtering to be a decode time operation we
missed a case that resulted in broken filtering. If we gave a list of
variants but did not include a scenario name the list would return
scenarnios with matching variants but would also include the scenario
name of scenarios that either didn't have a matrix defined or had no
matrix variants that matched.

Signed-off-by: Ryan Cragun <me@ryan.ec>