VAULT-41529: fix filtering by variant and no scenario name (#168)
* VAULT-41529: fix filtering by variant and no scenario name

When we rewrote scenario filtering to be a decode time operation, we missed a case that resulted in broken filtering in some cases. If we're given a filter that includes variants but does not include a scenario name, the list of scenarios with matching variants is returned but we also include the scenario name of scenarios that either didn't have a matrix defined or had no matrix variants that matched. This fixes it to exclude scenario blocks that have no matching variants when given such filters. Catching this bug earlier required fixing a test in sampling that expected an error much further along and also improving our earlier check.

* go: update modules
* remove old context package
* go mod tidy and add additional metadata to test

Signed-off-by: Ryan Cragun <me@ryan.ec>